+++ react/typings/gtm.d.ts
@@ -1,10 +1,12 @@
 export interface AnalyticsEcommerceProduct {
   id: string
-  skuId: string
   name: string
   category: string
   brand: string
   variant: string
   price: number
   quantity: number
+  dimension1: string
+  dimension2: string
+  dimension3: string
 }